Don't rely on canonical-link id, use [rel="canonical"]
This was relying on the canonical URL tag to have a "canonical-link" id, which
is only true for Guide pages. This now selects the <link> tag with the
rel="canonical" attribute, which applies for all pages.

This also removes some duplicated code by using has_bad_response_code in
has_url_mismatch instead of repeating the check for code['status'] == 200.